pyproject.toml
[project]
name = "pdf2audiobook"
version = "0.1.0"
description = "PDF to Audiobook SaaS Platform"
authors = [{name = "jpotterlabs", email = "jason@jpotterlabs.com"}]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "boto3>=1.34.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "paddle-sdk>=0.1.0",
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.10",
    "openai>=1.3.0",
    "pydub>=0.25.1",
    "reportlab>=4.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true